修改：修改了全局异常处理器以及注册拦截器的功能
-

+++ src/main/java/com/inabananami/japanesedemo/service/impl/WordServiceImpl.java
@@ -5,12 +5,14 @@ import com.inabananami.japanesedemo.dao.mapper.WordMapper;
 import com.inabananami.japanesedemo.dao.pojo.ReviewLog;
 import com.inabananami.japanesedemo.dao.pojo.Word;
 import com.inabananami.japanesedemo.service.WordService;
+import com.inabananami.japanesedemo.utils.ThreadLocalUtil;
 import com.inabananami.japanesedemo.vo.Result;
 import com.inabananami.japanesedemo.vo.WordVo;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Map;
 import java.util.stream.Collectors;
 
 @Service
@@ -22,6 +24,9 @@ public class WordServiceImpl implements WordService {
 
     @Override
     public Result add(Word word) {
+        Map<String,Object> map = ThreadLocalUtil.get();
+        Integer userId = (Integer) map.get("userId");
+        word.setCreateUser(userId);
         wordMapper.add(word);
         reviewMapper.defaultAdd(ReviewLog.defaultSet(word));
         return Result.success(null);
@@ -29,12 +34,29 @@ public class WordServiceImpl implements WordService {
 
     @Override
     public Result update(Word word) {
-        wordMapper.update(word);
+        Map<String,Object> map = ThreadLocalUtil.get();
+        Integer userId = (Integer) map.get("userId");
+        Word newWord = wordMapper.findWordById(word.getId());
+        if(newWord.getCreateUser() != userId){
+            return Result.fail(409, "此单词不是您的单词");
+        }
+        wordMapper.update(newWord);
         return Result.success(null);
     }
 
     @Override
     public Result delete(Integer id) {
+        Map<String,Object> map = ThreadLocalUtil.get();
+        Integer userId = (Integer) map.get("userId");
+        Word word = wordMapper.findWordById(id);
+        if(word == null){
+            return Result.fail(410, "此单词不存在");
+        }
+        ReviewLog reviewLog = reviewMapper.findReviewByWordId(word.getId());
+        if (word.getCreateUser() != userId) {
+            return Result.fail(409, "此单词不是您的单词");
+        }
+        reviewMapper.deleteReviewLog(reviewLog.getId());
         wordMapper.delete(id);
         return Result.success(null);
     }
@@ -51,5 +73,11 @@ public class WordServiceImpl implements WordService {
     @Override
     public Word findWordById(Integer id) {
         return wordMapper.findWordById(id);
+    }
+
+    @Override
+    public Result listAllWords() {
+        List<WordVo> wordList = wordMapper.listAll();
+        return Result.success(wordList);
     }
 }
